+++ spring-boot-chanalyzer/src/main/java/com/youtube/chanalyzer/dto/ChartJSDataResponseDTO.java
@@ -8,23 +8,23 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static com.youtube.chanalyzer.service.YTChannelDataService.sortVideosIntoMonths;
+import static com.youtube.chanalyzer.scraper.YouTubeChannelScraperAPI.sortVideosIntoMonths;
 
 
 @Data
 @NoArgsConstructor
-public class YTChannelDataResponseDTO {
+public class ChartJSDataResponseDTO {
     List<String> labels;
     List<Map<String, List<String>>> datasets;
     Integer currentInterval;
 
-    public YTChannelDataResponseDTO(ArrayList<HashMap<String, String>> responseObject) {
+    public ChartJSDataResponseDTO(ArrayList<HashMap<String, String>> responseObject) {
         var res = sortVideosIntoMonths(responseObject);
         this.labels = res.labels;
         this.datasets = res.datasets;
     }
 
-    public YTChannelDataResponseDTO setCurrentInterval(Integer currentInterval) {
+    public ChartJSDataResponseDTO setCurrentInterval(Integer currentInterval) {
         this.currentInterval = currentInterval;
         return this;
     }
