+++ spring-boot-chanalyzer/src/main/java/com/youtube/chanalyzer/scraper/ScraperAPI.java
@@ -1,0 +1,7 @@
+package com.youtube.chanalyzer.scraper;
+
+import reactor.core.publisher.Flux;
+
+public interface ScraperAPI {
+    Flux<?> getChannelVideoData(String url);
+}
